Make repo compatible as Ansible Role.
This change puts the `infoblox` module in a library directory so that it is compatible with Ansible role conventions.

The benefit of doing this is that once it is a role it is possible to specify a dependency on this role from any roles/requirements.yml file and have it be automatically downlaoded by Ansible Tower.